+++ app/not-found.tsx
@@ -3,14 +3,20 @@ import Link from "next/link";
 import { useTranslations } from "next-intl";
 import { getTranslations } from "next-intl/server";
 
-import { Mailto } from "./shared/mailto";
+import { Shell } from "./[locale]/layout/shell";
+import { Mailto } from "./[locale]/shared/mailto";
 
+/**
+ * Rendered for any pathname that does not match a page.
+ * This route is outside the [locale] segment, so it comes with its own HTML document.
+ * The locale is resolved by the request host (see i18n/request.ts).
+ */
 export default function NotFound() {
   const t = useTranslations("error");
   const tCommon = useTranslations("common");
 
   return (
-    <>
+    <Shell>
       <h1>{t("h1")}</h1>
       <p>
         {t.rich("explanation", {
@@ -23,7 +29,7 @@ export default function NotFound() {
       <div>
         <Link href="/">{tCommon("signature")}</Link>
       </div>
-    </>
+    </Shell>
   );
 }
